Remove OpDefTypeScheme, use PolyFuncType; move tests {type_scheme=>poly_func}.rs
