+++ src/extension.rs
@@ -26,7 +26,6 @@ pub use op_def::{CustomSignatureFunc, OpDef};
 mod type_def;
 pub use type_def::{TypeDef, TypeDefBound};
 pub mod prelude;
-pub mod type_scheme;
 pub mod validate;
 
 pub use prelude::{PRELUDE, PRELUDE_REGISTRY};
