--- src/extension/type_scheme.rs
@@ -1,271 +1,0 @@
-//! Polymorphic type schemes for [OpDef]s.
-//! The type scheme declares a number of TypeParams; any TypeArgs fitting those,
-//! produce a FunctionType for the Op by substitution.
-//!
-//! [OpDef]: super::OpDef
-
-use crate::types::type_param::{check_type_args, TypeArg, TypeParam};
-use crate::types::{FunctionType, Substitution};
-
-use super::{ExtensionRegistry, SignatureError};
-
-/// A polymorphic type scheme for an [OpDef]
-///
-/// [OpDef]: super::OpDef
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
-pub struct OpDefTypeScheme {
-    /// The declared type parameters, i.e., every instantiation ([ExternalOp]) must provide [TypeArg]s for these
-    ///
-    /// [ExternalOp]: crate::ops::custom::ExternalOp
-    pub params: Vec<TypeParam>,
-    /// Template for the Op type. May contain variables up to length of [OpDefTypeScheme::params]
-    body: FunctionType,
-}
-
-impl OpDefTypeScheme {
-    /// Create a new OpDefTypeScheme.
-    /// The [ExtensionRegistry] should be the same (or a subset) of that which will later
-    /// be used to create operations; at this point we only need the types.
-    ///
-    /// #Errors
-    /// Validates that all types in the schema are well-formed and all variables in the body
-    /// are declared with [TypeParam]s that guarantee they will fit.
-    pub fn new(
-        params: impl Into<Vec<TypeParam>>,
-        body: FunctionType,
-        extension_registry: &ExtensionRegistry,
-    ) -> Result<Self, SignatureError> {
-        let params = params.into();
-        body.validate(extension_registry, &params)?;
-        Ok(Self { params, body })
-    }
-
-    pub(super) fn compute_signature(
-        &self,
-        args: &[TypeArg],
-        extension_registry: &ExtensionRegistry,
-    ) -> Result<FunctionType, SignatureError> {
-        check_type_args(args, &self.params)?;
-        Ok(self
-            .body
-            .substitute(extension_registry, &Substitution::new(args, 0)))
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use std::num::NonZeroU64;
-
-    use smol_str::SmolStr;
-
-    use crate::extension::prelude::{PRELUDE_ID, USIZE_T};
-    use crate::extension::{
-        ExtensionId, ExtensionRegistry, SignatureError, TypeDefBound, TypeParametrised, PRELUDE,
-        PRELUDE_REGISTRY,
-    };
-    use crate::std_extensions::collections::{EXTENSION, LIST_TYPENAME};
-    use crate::types::type_param::{TypeArg, TypeArgError, TypeParam};
-    use crate::types::{CustomType, FunctionType, Type, TypeBound};
-    use crate::Extension;
-
-    use super::OpDefTypeScheme;
-
-    #[test]
-    fn test_opaque() -> Result<(), SignatureError> {
-        let list_def = EXTENSION.get_type(&LIST_TYPENAME).unwrap();
-        let tyvar = TypeArg::use_var(0, TypeParam::Type(TypeBound::Any));
-        let list_of_var = Type::new_extension(list_def.instantiate_concrete([tyvar.clone()])?);
-        let reg: ExtensionRegistry = [PRELUDE.to_owned(), EXTENSION.to_owned()].into();
-        let list_len = OpDefTypeScheme::new(
-            [TypeParam::Type(TypeBound::Any)],
-            FunctionType::new(vec![list_of_var], vec![USIZE_T]),
-            &reg,
-        )?;
-
-        let t = list_len.compute_signature(&[TypeArg::Type { ty: USIZE_T }], &reg)?;
-        assert_eq!(
-            t,
-            FunctionType::new(
-                vec![Type::new_extension(
-                    list_def
-                        .instantiate_concrete([TypeArg::Type { ty: USIZE_T }])
-                        .unwrap()
-                )],
-                vec![USIZE_T]
-            )
-        );
-
-        Ok(())
-    }
-
-    fn id_fn(t: Type) -> FunctionType {
-        FunctionType::new(vec![t.clone()], vec![t])
-    }
-
-    #[test]
-    fn test_mismatched_args() -> Result<(), SignatureError> {
-        let ar_def = PRELUDE.get_type("array").unwrap();
-        let typarams = [TypeParam::Type(TypeBound::Any), TypeParam::max_nat()];
-        let tyvar = TypeArg::use_var(0, typarams[0].clone());
-        let szvar = TypeArg::use_var(1, typarams[1].clone());
-
-        // Valid schema...
-        let good_array =
-            Type::new_extension(ar_def.instantiate_concrete([tyvar.clone(), szvar.clone()])?);
-        let good_ts = OpDefTypeScheme::new(typarams.clone(), id_fn(good_array), &PRELUDE_REGISTRY)?;
-
-        // Sanity check (good args)
-        good_ts.compute_signature(
-            &[TypeArg::Type { ty: USIZE_T }, TypeArg::BoundedNat { n: 5 }],
-            &PRELUDE_REGISTRY,
-        )?;
-
-        let wrong_args = good_ts.compute_signature(
-            &[TypeArg::BoundedNat { n: 5 }, TypeArg::Type { ty: USIZE_T }],
-            &PRELUDE_REGISTRY,
-        );
-        assert_eq!(
-            wrong_args,
-            Err(SignatureError::TypeArgMismatch(
-                TypeArgError::TypeMismatch {
-                    param: typarams[0].clone(),
-                    arg: TypeArg::BoundedNat { n: 5 }
-                }
-            ))
-        );
-
-        // (Try to) make a schema with bad args
-        let arg_err = SignatureError::TypeArgMismatch(TypeArgError::TypeMismatch {
-            param: typarams[0].clone(),
-            arg: szvar.clone(),
-        });
-        assert_eq!(
-            ar_def.instantiate_concrete([szvar.clone(), tyvar.clone()]),
-            Err(arg_err.clone())
-        );
-        // ok, so that doesn't work - well, it shouldn't! So let's say we just have this signature (with bad args)...
-        let bad_array = Type::new_extension(CustomType::new(
-            ar_def.name().clone(),
-            [szvar, tyvar],
-            PRELUDE_ID,
-            TypeBound::Any,
-        ));
-        let bad_ts = OpDefTypeScheme::new(typarams.clone(), id_fn(bad_array), &PRELUDE_REGISTRY);
-        assert_eq!(bad_ts.err(), Some(arg_err));
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_misused_variables() -> Result<(), SignatureError> {
-        // Variables in args have different bounds from variable declaration
-        let tv = TypeArg::use_var(0, TypeParam::Type(TypeBound::Copyable));
-        let list_def = EXTENSION.get_type(&LIST_TYPENAME).unwrap();
-        let body_type = id_fn(Type::new_extension(list_def.instantiate_concrete([tv])?));
-        let reg = [EXTENSION.to_owned()].into();
-        for decl in [
-            TypeParam::Extensions,
-            TypeParam::List(Box::new(TypeParam::max_nat())),
-            TypeParam::Type(TypeBound::Any),
-        ] {
-            let invalid_ts = OpDefTypeScheme::new([decl.clone()], body_type.clone(), &reg);
-            assert_eq!(
-                invalid_ts.err(),
-                Some(SignatureError::TypeVarDoesNotMatchDeclaration {
-                    used: TypeParam::Type(TypeBound::Copyable),
-                    decl: Some(decl)
-                })
-            );
-        }
-        // Variable not declared at all
-        let invalid_ts = OpDefTypeScheme::new([], body_type, &reg);
-        assert_eq!(
-            invalid_ts.err(),
-            Some(SignatureError::TypeVarDoesNotMatchDeclaration {
-                used: TypeParam::Type(TypeBound::Copyable),
-                decl: None
-            })
-        );
-
-        Ok(())
-    }
-
-    fn decl_accepts_rejects_var(
-        bound: TypeParam,
-        accepted: &[TypeParam],
-        rejected: &[TypeParam],
-    ) -> Result<(), SignatureError> {
-        const EXT_ID: ExtensionId = ExtensionId::new_unchecked("my_ext");
-        const TYPE_NAME: SmolStr = SmolStr::new_inline("MyType");
-
-        let mut e = Extension::new(EXT_ID);
-        e.add_type(
-            TYPE_NAME,
-            vec![bound.clone()],
-            "".into(),
-            TypeDefBound::Explicit(TypeBound::Any),
-        )
-        .unwrap();
-
-        let reg: ExtensionRegistry = [e].into();
-
-        let make_scheme = |tp: TypeParam| {
-            OpDefTypeScheme::new(
-                [tp.clone()],
-                id_fn(Type::new_extension(CustomType::new(
-                    TYPE_NAME,
-                    [TypeArg::use_var(0, tp)],
-                    EXT_ID,
-                    TypeBound::Any,
-                ))),
-                &reg,
-            )
-        };
-        for decl in accepted {
-            make_scheme(decl.clone())?;
-        }
-        for decl in rejected {
-            assert_eq!(
-                make_scheme(decl.clone()).err(),
-                Some(SignatureError::TypeArgMismatch(
-                    TypeArgError::TypeMismatch {
-                        param: bound.clone(),
-                        arg: TypeArg::use_var(0, decl.clone())
-                    }
-                ))
-            );
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_bound_covariance() -> Result<(), SignatureError> {
-        decl_accepts_rejects_var(
-            TypeParam::Type(TypeBound::Copyable),
-            &[
-                TypeParam::Type(TypeBound::Copyable),
-                TypeParam::Type(TypeBound::Eq),
-            ],
-            &[TypeParam::Type(TypeBound::Any)],
-        )?;
-
-        let list_of_tys = |b| TypeParam::List(Box::new(TypeParam::Type(b)));
-        decl_accepts_rejects_var(
-            list_of_tys(TypeBound::Copyable),
-            &[list_of_tys(TypeBound::Copyable), list_of_tys(TypeBound::Eq)],
-            &[list_of_tys(TypeBound::Any)],
-        )?;
-
-        decl_accepts_rejects_var(
-            TypeParam::max_nat(),
-            &[TypeParam::bounded_nat(NonZeroU64::new(5).unwrap())],
-            &[],
-        )?;
-        decl_accepts_rejects_var(
-            TypeParam::bounded_nat(NonZeroU64::new(10).unwrap()),
-            &[TypeParam::bounded_nat(NonZeroU64::new(5).unwrap())],
-            &[TypeParam::max_nat()],
-        )?;
-        Ok(())
-    }
-}
